+++ MotorcycleContest/repository/interfaces/IMotorcycleRepository.cs
@@ -3,7 +3,7 @@ using MotorcycleContest.model;
 
 namespace MotorcycleContest.repository.interfaces
 {
-    public interface IMotorcycleRepository : IRepository<long, Motorcycle>
+    public interface IMotorcycleRepository : IRepository<int, Motorcycle>
     {
         List<Motorcycle> FilterByEngineCapacity(EngineCapacity engineCapacity);
     }
