+++ MotorcycleContest/repository/interfaces/IRaceRepository.cs
@@ -3,10 +3,11 @@ using MotorcycleContest.model;
 
 namespace MotorcycleContest.repository.interfaces
 {
-    public interface IRaceRepository : IRepository<long, Race>
+    public interface IRaceRepository : IRepository<int, Race>
     {
         List<Race> FilterByParticipants(Participant participant);
         
         List<Race> FilterByRequiredEngineCapacity(EngineCapacity engineCapacity);
     }
 }
+
