+++ MotorcycleContest/repository/interfaces/IUserRepository.cs
@@ -3,7 +3,7 @@ using MotorcycleContest.model;
 
 namespace MotorcycleContest.repository.interfaces
 {
-    public interface IUserRepository : IRepository<long, User>
+    public interface IUserRepository : IRepository<int, User>
     {
         List<User> FilterByUsername(string username);
     }
